… go/src/knapsack/estimater.go
… go/src/knapsack/estimater_test.go