Use the incremental computation of usedCapacity and currentValue in computeEstimate instead of computing them every time.

Add a queue data structure.

**Benchmark before this change**
```
BenchmarkComputeOptimumKnapsack_ks_4_0-8    	  500000	      3508 ns/op	     968 B/op	      42 allocs/op
BenchmarkComputeOptimumKnapsack_ks_19_0-8   	   10000	    150317 ns/op	   43489 B/op	    1265 allocs/op
BenchmarkComputeOptimumKnapsack_ks_30_0-8   	     500	   3313536 ns/op	 1023896 B/op	   27339 allocs/op
BenchmarkComputeOptimumKnapsack_ks_50_0-8   	   10000	    120406 ns/op	   34304 B/op	    1004 allocs/op
BenchmarkComputeOptimumKnapsack_ks_200_0-8  	       1	10650347672 ns/op	2849250776 B/op	71286790 allocs/op
BenchmarkComputeOptimumKnapsack_ks_1000_0-8 	      50	  25180043 ns/op	10156703 B/op	   82846 allocs/op
BenchmarkComputeOptimumKnapsack_ks_10000_0-8	       2	 818236364 ns/op	77896488 B/op	  738968 allocs/op
```
**Benchmark after this change**
```
BenchmarkComputeOptimumKnapsack_ks_4_0-8    	 1000000	      1793 ns/op	     800 B/op	      24 allocs/op
BenchmarkComputeOptimumKnapsack_ks_19_0-8   	   30000	     44589 ns/op	   23512 B/op	     518 allocs/op
BenchmarkComputeOptimumKnapsack_ks_30_0-8   	    2000	   1031517 ns/op	  558453 B/op	   12683 allocs/op
BenchmarkComputeOptimumKnapsack_ks_50_0-8   	   50000	     32203 ns/op	   18248 B/op	     372 allocs/op
BenchmarkComputeOptimumKnapsack_ks_200_0-8  	       1	2065388807 ns/op	1201639592 B/op	24120411 allocs/op
BenchmarkComputeOptimumKnapsack_ks_1000_0-8 	     500	   2426590 ns/op	 1551455 B/op	   27887 allocs/op
BenchmarkComputeOptimumKnapsack_ks_10000_0-8	     100	  21922099 ns/op	17579887 B/op	  246999 allocs/op
```
For ks_200_0 solution the compute time drops from 10-11secs to 2-2.5secs.